Relocate css files into styles folder for readiness
